Cache Byte Buddy Proxy
For property rules that use a lambda expression we determine property
name by creating a byte buddy dynamic proxy and intercepting an invoke call
using the lambda expression. We were creating a new byte buddy proxy for
each rule which is inefficient  This commit caches dynamic proxy and the
internal property name interceptor.

Closes #11